requirements.txt
aiohttp==3.8.4
nextcord>=2.4.2
psycopg2-binary==2.9.6
pyspapi==2.1.2